# apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api Docs of E-Doc APP",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "tags": ["Review"],
        "summary": "Mendapatkan semua Review",
        "operationId": "getReviews",
        "responses": {
          "200": {
            "description": "Daftar Review berhasil diperoleh",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id_user": 1,
                    "id_content": 2,
                    "comment": "commant",
                    "createdAt": "2023-10-23"
                  },
                  {
                    "id_user": 2,
                    "id_content": 3,
                    "comment": "commant",
                    "createdAt": "2023-11-23"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Review"],
        "summary": "Menambahkan Review baru",
        "operationId": "insertReview",
        "requestBody": {
          "description": "Data yang dikirimkan",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_user": {"type": "integer"},
                  "id_content": {"type": "integer"},
                  "comment": {"type": "string"}
                }
              }
            },
            "application/json": {
              "example": {
                "id_user": "DataTypes.INTEGER",
                "id_content": "DataTypes.INTEGER",
                "comment": "DataTypes.String"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review berhasil ditambahkan",
            "content": {
              "application/json": {
                "example": { "status": true, "msg": "Data berhasil ditambahkan" }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{id}": {
      "get": {
        "tags": ["Review"],
        "summary": "Mendapatkan Review berdasarkan ID",
        "operationId": "getReviewById",
        "parameters": [
          {
            "name": "id",
            "in": "id",
            "description": "id review",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review berhasil ditemukan",
            "content": {
              "application/json": {
                "example": {
                  "id_review": 1,
                  "id_user": 2,
                  "id_content": 3,
                  "comment": "commant",
                  "createdAt": "2023-11-23"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": ["Review"],
        "summary": "Memperbarui Review berdasarkan ID",
        "operationId": "updateDataReview",
        "responses": {
          "200": {
            "description": "Review berhasil diperbarui",
            "content": {
              "application/json": {
                "example": { "status": true, "msg": "Rewview berhasil di perbarui" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Review"],
        "summary": "Menghapus Review berdasarkan ID",
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "id",
            "description": "id review",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review berhasil dihapus",
            "content": {
              "application/json": {
                "example": { "status": true, "msg": "Review berhasil dihapus" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/content-reaction-logs": {
      "post": {
        "tags": ["Content"],
        "summary": "Tambahkan reaction content",
        "operationId": "insertContentReactionLog",
        "requestBody": {
          "description": "Objek reaction log yang akan ditambahkan",
          "content": {
            "application/json": {
              "example": {
                "id_user": "DataTypes.INTEGER",
                "id_content": "DataTypes.INTEGER",
                "reaction": "DataTypes.BOOLEAN"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Log reaksi konten berhasil ditambahkan",
            "content": {
              "application/json": {
                "example": { "status": true }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/v1/content-reaction-logs/{id_content}": {
      "get": {
        "tags": ["Content"],
        "summary": "Dapatkan log reaksi konten berdasarkan ID Content",
        "operationId": "getContentReactionLogById",
        "parameters": [
          {
            "name": "id_content",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "example": {
                  "id_user": "12",
                  "id_content": "23",
                  "reaction": 1
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": ["Content"],
        "summary": "Perbarui log reaksi konten berdasarkan ID",
        "operationId": "updateContentReactionLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sukses",
            "content": {
              "application/json": {
                "example": { "status": true }
              }
            }
          },
          "500": {
            "description": "Error Server"
          }
        }
      }
    },
...
  }
}